ai-assistant/requirements.txt
transformers==4.35.2
torch==2.0.1
ollama==0.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
peft==0.7.1
bitsandbytes==0.41.2
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
sacrebleu==2.3.1
rouge-score==0.1.2
pillow==10.2.0
pytesseract==0.3.10
playwright==1.49.0
